requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
flask==3.0.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
openpyxl==3.1.2
fake-useragent==1.4.0
lxml==4.9.3
playwright==1.40.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
schedule==1.2.0
sqlalchemy==2.0.23
json5==0.9.14
retry==0.9.2
urllib3==2.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
Pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
